Render refactoring

+++ aui.views/src/AUI/Render/FontStyle.h
@@ -31,7 +31,7 @@ class AString;
 
 struct API_AUI_VIEWS FontStyle {
 	mutable _<AFont> font;
-	uint8_t size = 12;
+	unsigned size = 12;
 	bool formatting = false;
 	TextAlign align = TextAlign::LEFT;
 	AColor color;
